Revert "Add terrible hack to work around host/device issue"
This reverts commit 85ddf9f939730b68d6245ad28f917b3878b32788.
